+++ examples/ssr-hydrate-axum/src/app.rs
@@ -1,7 +1,7 @@
 use crate::error_template::{AppError, ErrorTemplate};
 use fluent_templates::static_loader;
 use leptos::*;
-use leptos_fluent::{i18n, leptos_fluent, move_tr, tr, Language};
+use leptos_fluent::{expect_i18n, leptos_fluent, move_tr, tr, Language};
 use leptos_meta::*;
 use leptos_router::*;
 
@@ -23,6 +23,7 @@ pub fn App() -> impl IntoView {
         initial_language_from_url_to_localstorage: true,
         initial_language_from_localstorage: true,
         initial_language_from_navigator: true,
+        set_to_localstorage: true,
         localstorage_key: "language",
         initial_language_from_accept_language_header: true,
     }};
@@ -52,14 +53,14 @@ pub fn App() -> impl IntoView {
 /// Renders the home page of your application.
 #[component]
 fn HomePage() -> impl IntoView {
-    let i18n = i18n();
+    let i18n = expect_i18n();
 
     view! {
         <h1>{move_tr!("welcome-to-leptos")}</h1>
         <fieldset>
             <For
                 each=move || i18n.languages
-                key=move |lang| format!("{}-{}", lang.id, **lang == i18n.language.get())
+                key=move |lang| i18n.language_key(lang)
                 children=move |lang: &&Language| {
                     view! {
                         <div>
@@ -68,8 +69,8 @@ fn HomePage() -> impl IntoView {
                                 id=lang.id.to_string()
                                 name="language"
                                 value=lang.id.to_string()
-                                checked=*lang == i18n.language.get()
-                                on:click=move |_| i18n.set_language_with_localstorage(lang)
+                                checked=i18n.is_active_language(lang)
+                                on:click=move |_| i18n.set_language(lang)
                             />
                             <label for=lang.id.to_string()>{lang.name}</label>
                         </div>