+++ leptos-fluent/src/localstorage.rs
@@ -1,7 +1,7 @@
 use cfg_if::cfg_if;
 
 pub fn get(#[allow(unused_variables)] key: &str) -> Option<String> {
-    cfg_if! { if #[cfg(all(not(feature = "ssr"), feature = "csr", feature = "hydrate"))] {
+    cfg_if! { if #[cfg(not(feature = "ssr"))] {
         ::leptos::window()
             .local_storage()
             .unwrap()
@@ -17,7 +17,7 @@ pub fn set(
     #[allow(unused_variables)] key: &str,
     #[allow(unused_variables)] value: &str,
 ) {
-    cfg_if! { if #[cfg(any(feature = "csr", feature = "hydrate"))] {
+    cfg_if! { if #[cfg(not(feature = "ssr"))] {
         ::leptos::window()
             .local_storage()
             .unwrap()