+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # CHANGELOG
 
+## 2024-05-20 - [0.0.24]
+
+### Enhancements
+
+- Add `I18n.is_active_language` method.
+- Add `I18n.language_key` method to return a hash for the current
+  language with their active status for usage in `For` components.
+- Add `set_to_localstorage` parameter to `leptos_fluent!` macro.
+- Add `use_i18n` and `expect_i18n` function.
+
+### Breaking changes
+
+- Replace `I18n.set_language_with_localstorage` method with
+  `I18n.set_language`. Use `set_to_localstorage` macro parameter
+  and `I18n.set_language` instead.
+- Remove `csr` feature.
+
+### Bug fixes
+
+- Fix errors getting initial language from URL parameter and local storage.
+
 ## 2024-05-18 - [0.0.23]
 
 - Add `axum` feature to integrate with Axum web framework.
@@ -44,6 +65,7 @@
 
 - Added all ISO-639-1 and ISO-639-2 languages.
 
+[0.0.24]: https://github.com/mondeja/leptos-fluent/compare/v0.0.23...v0.0.24
 [0.0.23]: https://github.com/mondeja/leptos-fluent/compare/v0.0.22...v0.0.23
 [0.0.22]: https://github.com/mondeja/leptos-fluent/compare/v0.0.21...v0.0.22
 [0.0.21]: https://github.com/mondeja/leptos-fluent/compare/v0.0.20...v0.0.21